Noise: Generate map as combination of octaves
The resultant noise is a combination of a number of octaves, whose
frequency and amplitude is determined by the lacuranity and persistance
provided by the user respectively. Generally, the frequency and
amplitude change via:

  frequency of an octave = lacuranity ^ octave_no
  amplitude of an octave = persistance ^ octave_no

The resultant map is a combination of all the octaves created in this
process.

Each octave is generated via an octave offset, which is generated pseudo
randomly depending on the value of a "seed" provided, so that we can get
a specific map using a specific seed.

"halfWidth" and "halfHeight" is used to determine the mid poin position
of the map, so that the "zooming" while moving the noiseScale slider is
done along the center and not along the top right corner.

The noiseMap[,] is clamped in the range of [0,1] later using
Mathf.InverseLerp().

Finally, in the MapGenerator.cs, we provide a Tooltip for few of the
public variables, and limit it's value using the "OnValidate()" method.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>