Mesh: Generate Mesh from Noise Map
Added a new DrawMode in MapGenerator.cs named "Mesh" which will create a
mesh representing the terrain generated via the noise map.

The mesh is defined in the MeshGenerator.cs script, which takes in the
heightmap as input and defines the mesh using this data.

Finally, the mesh is drawn in the MapDisplay class, in the DrawMesh()
function by calling the meshData.CreateMesh() function

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>