LOD: Added LOD and height multiplier in mesh
The "y" coordinate of the mesh is multiplied by "heightMultiplier" to
adjust the height of the mountains. Also, we want the water levels to be
flat, which is why the y coordinate of the mesh is set as:

  heightCurve.Evaluate(heightMap[x, y]) * heightMultiplier

where heightCurve is an AnimationCurve with exponential rise for x > 0.4
and y = 0 for x <= 0.4, thus making the water level flat.

Also added levelOfDetail which will be used later to optimise the mesh
present further away from the camera. On increasing the levelOfDetail,
we are "skipping" few elements on both the width and the height of the
mesh, thus making the mesh less detailed.

We choose a mapChunkSize of 241 on both width and height, since we will
get a proper number of vertices for LOD = [1, 6]. The number of vertices
along a given line is given as:

  verticesPerLine = (width - 1) / increment + 1

Here increment = 2, 4, 6, 8, 10, 12 and width = 241, thus making it
divisible.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>