LOD switching, collision of the terrain mesh
Implemented LOD switching to switch LOD of the meshes depending on how
close the mesh is from the player. The terrain is made seamless by
calculating the value of the normals of one terrain and equating it with
the value of the normals of the next terrain. The two terrains of
different LOD levels are not seamless, but it won't be a problem since
the user won't be able to view it.

Implemented collision of the terrain mesh by adding MeshCollider to the
mesh. For optimisation, the collision is added only when it is needed
and not every time they are generated.

Signed-off-by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>